Adds comments throughout debbuild script
Took a pass through on cleanup, based on `shellcheck` output. Mostly
documented the script so that future maintainers can easily pick up and
understand what the script is doing.

Also added some automagic, in that the build script will look for a
suitable tarball in an adjacent directory. Assuming developers have just
built the tarball, as the build script requires them to do, this is a
reasonable assumption.

Even better would be to fetch the remote, verify the tag, and build the
tarball locally. We can add that functionality later.