Resolves venv paths in python deb packages
The new build logic that came in via #9 introduced a regression, where
the venv path inside the scripts packaged in the debs was
inappropriately using a fullpath from the build environment.

Modifying the pardir to match the package name exactly resolved; it was
necessary to tweak the tar command to make sure the files land adjacent
to the debian dir as the build tooling expects.

Closes #12.