Renames syncwheels -> fetch-wheels
The operation is strictly one-way, downloading the remote wheels from S3
and storing them in the `localwheels/` dir inside the packaging
repository. For clarity, renaming the action to distinguish it from the
*other* one-way action, pushing new wheels up to S3 (not yet codified).

Also implements an optimization on the fetching operation, via the
`--continue-at` flag to curl. Essentially the files will not be
redownloaded if they exist locally. It's OK to skip a full download,
since we'll be verifying integrity of the contents of the S3 buckets via
a SHA256SUMS file, with detached sig, inside this packaging repo.