Adds script for verifying sig on checksums
The sha256sums file is used to verify the integrity of assets pulled in
from S3; in order to trust the contents of the sha256sums file itself,
we'll verify it via a detached, ASCII-armored GPG signature, stored in
this packaging repository.

We're creating a temporary keyring to perform the verification,
otherwise gpg will report success if *any* key in the local keyring was
used to create the detached sig; that's not what we want. Instead, only
a small list of pubkeys should be considered valid, so we'll store those
pubkeys here.

Sprinkles the new script logic in a few locations, where we want to
check the integrity; pending review, we may revise the locations of
these calls out to the detached sig verification.