Adds git-lfs to installdeps target
Provides a convenient method for installing git-lfs, as part of the
install-required-dependencies step. Added a conditional check to ensure
that the wheels are indeed retrieved via git-lfs. This change is
required by developers only on first-run, thus the conditional, but
required in CI all the time, since the repo is cloned before deps are
installed.