CI: Add securedrop-workstation-config metapackage ci job
This package is only used for the Template build process, and for now only used to pull in dependencies. Arguably could be simplified by explicitly installing as part of the template install script, but could be helpful in the future.

Omitting nightlies of this package for simplicity, we should revisit if we make broader changes to this package.