packaging scripts: remove wheelsurls.txt and related scripts
We used to keep track of the URLs of all wheels such that we
could easily download them all when we were using S3. Now we
store the wheels in this git-lfs repository, so there is no
need to sync all wheels as it is done automatically by resolving
the git-lfs pointers.