Adds error message if tag doesn't verify
Requested by @kushaldas during review. We can't easily pass through the stderr
from the git verify process, since we redirect that stderr to stdout so
that grep can inspect it. So let's just detect the failure and report
that the tag failed to verify, which is accurate, albeit neither
explicit nor verbose.