Checks for u2mfn kernel module in metapackage
The u2mfn kernel module is required for GUI operations in a VM. It
should be built automatically via dkms, but that process can fail. We
must report such errors up to the parent apt operation to notify users.
The /etc/kernel/postinst.d/dkms hooks are run on installation of the
linux-image-* packages, so we don't need an explicit call to dkms
autoinstall (which fails silently) in the metapackage postinst.

Since we've observed VMs fail to build the u2mfn.ko dynamically in the
past, let's try to recover in that situation, otherwise fail loudly.