open-in-dvm.desktop already supplied by other packages
Current versions of securedrop-client, securedrop-export, securedrop-workstation-svs-disp already manage this file in /usr/share/applications/, which causes an error on install. In order to allow the coexistance of these packages, we need to place and move these files manually ourselves. Once we clean up the other packages (at the final phase of template-consolidation) we can manage this file properly directly through the .install file, and have it handled by this package. All other packages (client, export, svs-disp) will not handle the mimetype config, it will be handled via salt in AppVM's private volume.