Adds script to generate tag message
When tagging the "securedrop-debian-packaging" repository, we include
a list of all package versions, to make reasoning about releases easier
down the road. Collecting this information manually is error-prone, so
let's provide a routine method of generating that info so it can easily
be pasted into a prompt for an annotated git tag, consistent with docs.