Cleans up "default" option for mimetypes in config pkg
Pointed out by @emkll during review, the ".install" declarations use
directories as dest, so we can't write a file that way. Added to
postinst, along with the other setups.

Technically we could reuse the ".sd-app" mimetype config elsewhere, but
using a ".default" name for it keeps the door open on divergence in the
future, and matches what we've been testing as part of template
consolidation.