Drops custom PyPI for securedrop-export
Updated the build logic for securedrop-export to remove use of the
https://pypi.securedrop.org URL. At present, the securedrop-export
doesn't use any external Python dependencies, it's core-only. So the
custom PyPI URL was never necessary, and technically we could excise it
altogether right now, but for consistency's sake I'm updating it to
match the other Python-based SDW packages.