Uses system python for wheels building
Now we are using the Debian provided Python to create the
virtualenv. Means the wheels will also be created using the system
python, and will have the exact same sha256sums.

By default the CircleCI images are based on the docker Python images
and they install python from building the source.