Update proxy reqs path
Recent nightlies of `securedrop-proxy` were broken, due to missing
imports [0]. In a recent PR [1], the requirements files within the
securedrop-proxy repo (and therefore tarball) were moved, and the
package build logic depends on those locations.

To my great surprise and disappoint, dh_virtualenv evidently will not
complain if a `--requirements` file does not exist. The build will be
called successful. That's unacceptable, so I've modified the
debian/rules file to test for existence as a workaround. If folks agree
this is a sane solution, we should update the rules files for the other
packages to match.

[0] https://github.com/freedomofpress/securedrop-proxy/issues/94
[1] https://github.com/freedomofpress/securedrop-proxy/pull/92