Fix inferring version in build pkg script
The upstream raw html on GitHub release pages has changed, and doesn't
include the tag info we want. That's fine: we'll just query tags
directly. The function fix here is the "releases" -> "tags" change in
the curl call; the rest of the diff is the result of running shellcheck
against the script and addressing the problems reported.

Closes #277.