Build native packages with debian/ in source tree
The end goal is that each repository has all the necessary files to
build a Debian package, without needing this repository. This gets us
partially there for metadata in the debian/ directory.

Currently we build a sdist tarball and use it as the source tree for
packaging. This has some nice properties in that we can exclude some
unnecessary files and lets us control timestamps for reproducibility.
This comes at a higher cognitive cost of needing to know everything that
ends up in a sdist tarball and complexity to reset timestamps.

With this change, we now use the Git repo tree as the source tree for
packaging. Steps like copying over a debian/ directory from this
repository and installing a platform-specific changelog are guarded on
files existing so they can be removed/adjusted on a repo-by-repo basis.

Refs #308.