Add a bullseye builder
Duplicate the current dockerfiles/ directory into a separate buster
and bullseye directories with the understanding that the buster
one is going to go away pretty soon, so it's not worth abstracting
the logic to be platform agnostic.

I dropped some packages and the timezone step that were already present
in the base `debian:bullseye` image.