Fix building packages when PKG_PATH doesn't end in PKG_NAME
There's a subtle bug in which we (read: I) incorrectly assumed that
PKG_PATH will always end in PKG_NAME, so copying it over into
$TOP_BUILDDIR will create the appropriately named subdirectory and
everything is happy.

This assumption is wrong in all the component repository CI pipelines,
where the path points to a directory named "project". Adjust the `cp`
command with the magic `-T` flag to stick the files in the correctly
named directory that we expect later on.

Thanks to Gonzalo for helping with debugging.