Parallelize nightly builds
Instead of having each nightly job push to the apt-test repo, they now
store the resulting .deb in a CircleCI workspace[1], which the final
"push-packages" job will copy from and commit.

This provides a number of benefits, including:
* faster nightlies since builds are parallelized
* no maintaining a complex "requires" list of jobs
* build jobs don't have access to SSH key for apt-test
* only two commits instead of 10-12

It also unlocks future improvements, such as:
* consolidating the build- and build-nightly jobs for DRY
* folding cleanup-nightlies script into the same push operation

Finally I expect this (and future DRY work) to make it easier to add
bookworm support in the near future.

[1] https://circleci.com/docs/workspaces