Switch to plain Debian images in CI
For the most part this is a very straightforward switch, except for the
reprotest-wheels job.

As noted in the previous commit, we cannot run the build-sync-wheels
script as root (which the Debian images default to). We add a "ci" user,
chown the repository for them, and then run the test command. One subtle
catch is that the repository needs to be checked out in a place other
than /root, so we set the working_directory to /srv.

I also noticed that we were installing all of the 5GB of diffoscope
dependencies without actually ever using diffoscope (and if we were
using diffoscope, we don't need extra utilities to diff debs and
wheels), so using `--no-install-recommends` shaves off about 3 minutes
of runtime. We do need to manually install faketime since it's only a
recommends, but that seems like a fair tradeoff.

And we can now delete the custom `packaging-debian-{buster,bullseye}`
images we were maintaining!