Remove buster, add failing bookworm jobs
We'd like bookworm jobs to be able to fail, but when they do fail it
should make an appropriate level of noise so we can investigate why they
failed. But they should also not stop the bullseye packages from being
uploaded!

So we split the bookworm jobs into a separate track ("build2"), and have
separate push jobs for each version. The push-bullseye job goes before
the push-bookworm one in case bookworm jobs are failing.

Note that currently all the bookworm jobs are failing because of
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1017046. I've put in a
fix for that, it should be in bookworm in the next 3-5 days.

Refs #369.