Apply a consistent sort to sha256sums.txt files
The ordering of `sha256sum *` is not consistent, it depends on specifics
of the shell and filesystem ordering. This causes unnecessary diffs when
regenerating the files.

Have Python sort the output and save it to ensure a consistent
file order. I opted for a case-sensitive comparison since that's what
most of the files were already using. The only unsorted file is
securedrop-client, which is modified in a follow-up.