Add empty sha256sums.txt files for securedrop-export
securedrop-export currently doesn't have any dependencies, but we treat
it like it does. However, it didn't have a wheels folder, nor
sha256sums.txt{,.asc}, which broke some of the scripts.

The scripts are now tolerant of empty wheels directories and will
generate a sha256sums.txt with "# No files" in it instead of a blank
file.

Fixes #394.