Refactor clean-old-nightlies so it works for all package folders
The cleanup script can now be used for the non-nightly directories too.
Instead of trying to parse a timestamp out of the filename, it sorts the
packages by version and only keeps the specified number.

It uses the proper `dpkg` version comparing function, and properly
supports the Linux packages that have a different naming scheme.