Automatically clean up non-nightly packages
Instead of relying on humans to delete the older packages when they add
new ones, let's automate it.

CI will now keep the 4 latest versions of a package, deleting the rest
as part of the nightly run.

While we're at it, lower the number of nightlies to be kept from 14 days
to 7 days. The 14 days amount was rather conservative when we first
implemented automation, we can probably be slightly more aggressive now.