Add script to simplify promoting packages from dev to prod
The new `./scripts/promote-suite` simplifies the process of promoting
packages from the dev repository to the prod one.

The primary use case currently is for promoting Tor and Linux packages.

For each package in the dev repo, it compares the highest version to the
highest version in the prod repo, copying over the dev package if they
don't match. There are some edge cases where this doesn't work (which is
annotated with a code comment), but it should get the majority correct.

As a side-effect, it could reveal when a package in dev was forgotten
for promotion to prod.

In the future this script could be fancier by showing debdiffs and
diffoscopes of the newly promoted packages, or a dry-run mode that shows
which packages are pending promotion.