Have CI build nightlies for securedrop-updater
The job increments the version number in the specfile,
runs `make rpm-build` and commits the result to the
securedrop-workstation-dev-rpm-packages-lfs repository.

For now this is specific to securedrop-updater, in the future
we expect to use the same build commands for securedrop-workstation
and will set up nightlies for that too.

Fixes <https://github.com/freedomofpress/securedrop-updater/issues/10>.