Create issues when a new Tor release is fetched
CI will now create a new issue (or update an existing one) whenever it
fetches new Tor packages. This is something that I have been doing
manually for a while now whenever I see a new Tor release announcement.

The generated issue contains the checklist as well as the diff of the
new debs so you can see the version and checksums of the packages. An
example issue (with the wrong patch) is
<https://github.com/freedomofpress/securedrop/issues/6723>.

Some potential future improvements that I deferred for now is to include
the version number in the issue title, and the correct Tor Project forum
link.

The script wraps around the official `gh` CLI tool, it needs a
GITHUB_TOKEN to be set in the environment to work properly. `gh` is only
available in bullseye-backports, so I had to adjust the image so it
would be installable.

Of course, if this works out well, I'd like to expand this to other
things like kernel and dependency updates.