Rename "PKG_GITREF" to "SD_PKG_GITREF"
This will be exported in the `.buildinfo` file so let's prefix it with
"SD" to make it abundantly clear this is our environment variable and
reduce the risk of collision.

There don't appear to be any other users of this besides humans, so I
have not added in any backwards-compatibility support for the old name.