Add SecureDrop-specific metadata to buildinfo files
dpkg will only export specific known environment variables into the
buildinfo file[1], but we want to track at least two more things:

 1) the Git commit of the package being built
 2) the Git commit of the securedrop-builder repository being used, as
    the bootstrap and wheels used affect the package output.

We can track those by capturing the values before the build process and
then manually adding them to the end of the buildinfo file. This also
means that builds must be done from a Git checkout, and cannot be from a
tarball; so that path now errors out.

It would be too complicated to have this script respect the new
$SD_BUILDER_GITREF by checking out a different version of the
repository, re-initializing the bootstrap, etc., so we leave that as a
responsiblity of the caller. But as a failsafe we still double-check
that the desired version of the builder repository is being used.

While we're at it, look up the package filename by "parsing"
debian/files instead of trying to find it via find.

[1]
https://git.dpkg.org/cgit/dpkg/dpkg.git/tree/scripts/Dpkg/BuildInfo.pm