Fix issues around case-sensitive matching in poetry support
poetry doesn't normalize case in package names, so in securedrop-client
we have jinja2 depending on "MarkupSafe", but the package itself in the
lock file is named "markupsafe".

We can avoid this by changing the relevant_dependencies to all be
lowercase when reading them out of pyproject.toml (the "main"
dependencies) and when reading from poetry.lock.