Cleanup old RPM nightlies too
Extend the `clean-old-packages` script so it works for RPMs too. We use
the Python RPM bindings to extract package name and version and then
shell out to `rpmdev-vercmp` to sort them.

The behavior is now the same as the deb cleanup, when new nightly
packages are pushed, it'll clean up old nightlies in the same commit. RC
packages (in the `f32` component) will be cleaned up in a separate
commit, if necessary.

Fixes #435.