Remove files related to Debian packaging, now in client repo
All of the packaging logic has moved to the securedrop-client
repository, leaving this responsible for the localwheels building. (RPM
nightlies will be moved out at a later time.)

Nearly everything has been migrated, the things that haven't
(lint-desktop-files, test_deb_package.py and clean-old-packages) have
open issues/PRs for them.

One explicit workflow change is that PRs to this repo will not test
package builds, because doing so causes a catch-22 scenario when the
build-requirements.txt in the client repo hasn't been updated yet, so it
fails.